Revert "Pull request inspection and corrections"
