+++ packages/smooth_app/lib/bottom_sheet_views/user_contribution_view.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:smooth_app/functions/launchURL.dart';
 import 'package:smooth_ui_library/buttons/smooth_simple_button.dart';
-import 'package:smooth_ui_library/dialogs/smooth_alert_dialog.dart';
+import 'package:smooth_ui_library/dialogs/smooth_AlertDialog.dart';
 import 'package:smooth_ui_library/widgets/smooth_listTile.dart';
 
 Launcher launcher = Launcher();
@@ -102,7 +102,7 @@ class UserContributionView extends StatelessWidget {
         context: context,
         builder: (BuildContext context) {
           return SmoothAlertDialog(
-            //context: context,
+            context: context,
             close: false,
             title: AppLocalizations.of(context).contribute_contribute_header,
             body: Column(
@@ -163,7 +163,7 @@ class UserContributionView extends StatelessWidget {
       builder: (BuildContext context) {
         return SmoothAlertDialog(
           title: AppLocalizations.of(context).contribute_develop,
-          //context: context,
+          context: context,
           body: Column(
             children: <Widget>[
               Text(
@@ -216,7 +216,7 @@ class UserContributionView extends StatelessWidget {
       builder: (BuildContext context) {
         return SmoothAlertDialog(
           title: AppLocalizations.of(context).contribute_translate_header,
-          //context: context,
+          context: context,
           body: Column(
             children: <Widget>[
               Text(
@@ -247,7 +247,7 @@ class UserContributionView extends StatelessWidget {
       context: context,
       builder: (BuildContext context) {
         return SmoothAlertDialog(
-          //context: context,
+          context: context,
           body: Column(
             children: <Widget>[
               Text(
